+++ src/oci/core/models/dedicated_vm_host.py
@@ -9,7 +9,8 @@ from oci.decorators import init_model_state_from_kwargs
 @init_model_state_from_kwargs
 class DedicatedVmHost(object):
     """
-    A dedicated virtual machine (VM) host that enables you to host multiple virtual machine instances on a dedicated host that is not shared with other tenancies.
+    A dedicated virtual machine host that enables you to host multiple VM instances
+    on a dedicated host that is not shared with other tenancies.
     """
 
     #: A constant which can be used with the lifecycle_state property of a DedicatedVmHost.
@@ -139,7 +140,7 @@ class DedicatedVmHost(object):
     def availability_domain(self):
         """
         **[Required]** Gets the availability_domain of this DedicatedVmHost.
-        The availability domain the dedicated VM host is running in.
+        The availability domain the dedicated virtual machine host is running in.
 
         Example: `Uocm:PHX-AD-1`
 
@@ -153,7 +154,7 @@ class DedicatedVmHost(object):
     def availability_domain(self, availability_domain):
         """
         Sets the availability_domain of this DedicatedVmHost.
-        The availability domain the dedicated VM host is running in.
+        The availability domain the dedicated virtual machine host is running in.
 
         Example: `Uocm:PHX-AD-1`
 
@@ -167,7 +168,7 @@ class DedicatedVmHost(object):
     def compartment_id(self):
         """
         **[Required]** Gets the compartment_id of this DedicatedVmHost.
-        The OCID of the compartment that contains the dedicated VM host.
+        The OCID of the compartment that contains the dedicated virtual machine host.
 
 
         :return: The compartment_id of this DedicatedVmHost.
@@ -179,7 +180,7 @@ class DedicatedVmHost(object):
     def compartment_id(self, compartment_id):
         """
         Sets the compartment_id of this DedicatedVmHost.
-        The OCID of the compartment that contains the dedicated VM host.
+        The OCID of the compartment that contains the dedicated virtual machine host.
 
 
         :param compartment_id: The compartment_id of this DedicatedVmHost.
@@ -191,7 +192,7 @@ class DedicatedVmHost(object):
     def dedicated_vm_host_shape(self):
         """
         **[Required]** Gets the dedicated_vm_host_shape of this DedicatedVmHost.
-        The shape of the dedicated VM host. The shape determines the number of CPUs and
+        The dedicated virtual machine host shape. The shape determines the number of CPUs and
         other resources available for VMs.
 
 
@@ -204,7 +205,7 @@ class DedicatedVmHost(object):
     def dedicated_vm_host_shape(self, dedicated_vm_host_shape):
         """
         Sets the dedicated_vm_host_shape of this DedicatedVmHost.
-        The shape of the dedicated VM host. The shape determines the number of CPUs and
+        The dedicated virtual machine host shape. The shape determines the number of CPUs and
         other resources available for VMs.
 
 
@@ -281,14 +282,18 @@ class DedicatedVmHost(object):
     def fault_domain(self):
         """
         Gets the fault_domain of this DedicatedVmHost.
-        The fault domain for the dedicated VM host's assigned instances. For more information, see Fault Domains.
+        The fault domain for the dedicated virtual machine host's assigned instances.
+        For more information, see `Fault Domains`__.
 
-        If you do not specify the fault domain, the system selects one for you. To change the fault domain for a dedicated VM host,
-        delete it and create a new dedicated VM host in the preferred fault domain.
+        If you do not specify the fault domain, the system selects one for you. To change the fault domain for a dedicated virtual machine host,
+        delete it, and then create a new dedicated virtual machine host in the preferred fault domain.
 
-        To get a list of fault domains, use the ListFaultDomains operation in the Identity and Access Management Service API.
+        To get a list of fault domains, use the `ListFaultDomains` operation in the `Identity and Access Management Service API`__.
 
         Example: `FAULT-DOMAIN-1`
+
+        __ https://docs.cloud.oracle.com/iaas/Content/General/Concepts/regions.htm#fault
+        __ https://docs.cloud.oracle.com/iaas/api/#/en/identity/20160918/
 
 
         :return: The fault_domain of this DedicatedVmHost.
@@ -300,14 +305,18 @@ class DedicatedVmHost(object):
     def fault_domain(self, fault_domain):
         """
         Sets the fault_domain of this DedicatedVmHost.
-        The fault domain for the dedicated VM host's assigned instances. For more information, see Fault Domains.
+        The fault domain for the dedicated virtual machine host's assigned instances.
+        For more information, see `Fault Domains`__.
 
-        If you do not specify the fault domain, the system selects one for you. To change the fault domain for a dedicated VM host,
-        delete it and create a new dedicated VM host in the preferred fault domain.
+        If you do not specify the fault domain, the system selects one for you. To change the fault domain for a dedicated virtual machine host,
+        delete it, and then create a new dedicated virtual machine host in the preferred fault domain.
 
-        To get a list of fault domains, use the ListFaultDomains operation in the Identity and Access Management Service API.
+        To get a list of fault domains, use the `ListFaultDomains` operation in the `Identity and Access Management Service API`__.
 
         Example: `FAULT-DOMAIN-1`
+
+        __ https://docs.cloud.oracle.com/iaas/Content/General/Concepts/regions.htm#fault
+        __ https://docs.cloud.oracle.com/iaas/api/#/en/identity/20160918/
 
 
         :param fault_domain: The fault_domain of this DedicatedVmHost.