+++ src/oci/core/models/ip_sec_connection.py
@@ -479,8 +479,12 @@ class IPSecConnection(object):
         you must provide at least one valid static route. If you configure both
         tunnels to use BGP dynamic routing, you can provide an empty list for the static routes.
 
+        The CIDR can be either IPv4 or IPv6. Note that IPv6 addressing is currently supported only
+        in the Government Cloud.
 
         Example: `10.0.1.0/24`
+
+        Example: `2001:db8::/32`
 
 
         :return: The static_routes of this IPSecConnection.
@@ -500,8 +504,12 @@ class IPSecConnection(object):
         you must provide at least one valid static route. If you configure both
         tunnels to use BGP dynamic routing, you can provide an empty list for the static routes.
 
+        The CIDR can be either IPv4 or IPv6. Note that IPv6 addressing is currently supported only
+        in the Government Cloud.
 
         Example: `10.0.1.0/24`
+
+        Example: `2001:db8::/32`
 
 
         :param static_routes: The static_routes of this IPSecConnection.