+++ src/oci/identity/models/create_region_subscription_details.py
@@ -43,6 +43,9 @@ class CreateRegionSubscriptionDetails(object):
         - `IAD`
         - `FRA`
         - `LHR`
+        - `YYZ`
+        - `NRT`
+        - `ICN`
 
         Example: `PHX`
 
@@ -63,6 +66,9 @@ class CreateRegionSubscriptionDetails(object):
         - `IAD`
         - `FRA`
         - `LHR`
+        - `YYZ`
+        - `NRT`
+        - `ICN`
 
         Example: `PHX`