+++ src/oci/identity/models/policy.py
@@ -23,6 +23,9 @@ class Policy(object):
     To use any of the API operations, you must be authorized in an IAM policy. If you're not authorized,
     talk to an administrator.
 
+    **Warning:** Oracle recommends that you avoid using any confidential information when you supply string values
+    using the API.
+
     __ https://docs.cloud.oracle.com/Content/Identity/Concepts/overview.htm
     __ https://docs.cloud.oracle.com/Content/Identity/Concepts/policygetstarted.htm
     """