+++ src/oci/identity/models/tag_default_summary.py
@@ -55,6 +55,10 @@ class TagDefaultSummary(object):
             Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
         :type lifecycle_state: str
 
+        :param is_required:
+            The value to assign to the is_required property of this TagDefaultSummary.
+        :type is_required: bool
+
         """
         self.swagger_types = {
             'id': 'str',
@@ -64,7 +68,8 @@ class TagDefaultSummary(object):
             'tag_definition_name': 'str',
             'value': 'str',
             'time_created': 'datetime',
-            'lifecycle_state': 'str'
+            'lifecycle_state': 'str',
+            'is_required': 'bool'
         }
 
         self.attribute_map = {
@@ -75,7 +80,8 @@ class TagDefaultSummary(object):
             'tag_definition_name': 'tagDefinitionName',
             'value': 'value',
             'time_created': 'timeCreated',
-            'lifecycle_state': 'lifecycleState'
+            'lifecycle_state': 'lifecycleState',
+            'is_required': 'isRequired'
         }
 
         self._id = None
@@ -86,6 +92,7 @@ class TagDefaultSummary(object):
         self._value = None
         self._time_created = None
         self._lifecycle_state = None
+        self._is_required = None
 
     @property
     def id(self):
@@ -288,6 +295,44 @@ class TagDefaultSummary(object):
         if not value_allowed_none_or_none_sentinel(lifecycle_state, allowed_values):
             lifecycle_state = 'UNKNOWN_ENUM_VALUE'
         self._lifecycle_state = lifecycle_state
+
+    @property
+    def is_required(self):
+        """
+        **[Required]** Gets the is_required of this TagDefaultSummary.
+        If you specify that a value is required, a value is set during resource creation (either by
+        the user creating the resource or another tag defualt). If no value is set, resource
+        creation is blocked.
+
+        * If the `isRequired` flag is set to \"true\", the value is set during resource creation.
+        * If the `isRequired` flag is set to \"false\", the value you enter is set during resource creation.
+
+        Example: `false`
+
+
+        :return: The is_required of this TagDefaultSummary.
+        :rtype: bool
+        """
+        return self._is_required
+
+    @is_required.setter
+    def is_required(self, is_required):
+        """
+        Sets the is_required of this TagDefaultSummary.
+        If you specify that a value is required, a value is set during resource creation (either by
+        the user creating the resource or another tag defualt). If no value is set, resource
+        creation is blocked.
+
+        * If the `isRequired` flag is set to \"true\", the value is set during resource creation.
+        * If the `isRequired` flag is set to \"false\", the value you enter is set during resource creation.
+
+        Example: `false`
+
+
+        :param is_required: The is_required of this TagDefaultSummary.
+        :type: bool
+        """
+        self._is_required = is_required
 
     def __repr__(self):
         return formatted_flat_dict(self)