+++ src/oci/identity/models/update_tag_default_details.py
@@ -21,16 +21,23 @@ class UpdateTagDefaultDetails(object):
             The value to assign to the value property of this UpdateTagDefaultDetails.
         :type value: str
 
+        :param is_required:
+            The value to assign to the is_required property of this UpdateTagDefaultDetails.
+        :type is_required: bool
+
         """
         self.swagger_types = {
-            'value': 'str'
+            'value': 'str',
+            'is_required': 'bool'
         }
 
         self.attribute_map = {
-            'value': 'value'
+            'value': 'value',
+            'is_required': 'isRequired'
         }
 
         self._value = None
+        self._is_required = None
 
     @property
     def value(self):
@@ -55,6 +62,44 @@ class UpdateTagDefaultDetails(object):
         :type: str
         """
         self._value = value
+
+    @property
+    def is_required(self):
+        """
+        Gets the is_required of this UpdateTagDefaultDetails.
+        If you specify that a value is required, a value is set during resource creation (either by
+        the user creating the resource or another tag defualt). If no value is set, resource
+        creation is blocked.
+
+        * If the `isRequired` flag is set to \"true\", the value is set during resource creation.
+        * If the `isRequired` flag is set to \"false\", the value you enter is set during resource creation.
+
+        Example: `false`
+
+
+        :return: The is_required of this UpdateTagDefaultDetails.
+        :rtype: bool
+        """
+        return self._is_required
+
+    @is_required.setter
+    def is_required(self, is_required):
+        """
+        Sets the is_required of this UpdateTagDefaultDetails.
+        If you specify that a value is required, a value is set during resource creation (either by
+        the user creating the resource or another tag defualt). If no value is set, resource
+        creation is blocked.
+
+        * If the `isRequired` flag is set to \"true\", the value is set during resource creation.
+        * If the `isRequired` flag is set to \"false\", the value you enter is set during resource creation.
+
+        Example: `false`
+
+
+        :param is_required: The is_required of this UpdateTagDefaultDetails.
+        :type: bool
+        """
+        self._is_required = is_required
 
     def __repr__(self):
         return formatted_flat_dict(self)